[REFACTOR] move upgrade profile to /auth
